update error report + add flag reco

+++ srcs/ft_nm/handle_x32.c
@@ -6,7 +6,7 @@
 /*   By: ddinaut <ddinaut@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/26 14:33:41 by ddinaut           #+#    #+#             */
-/*   Updated: 2018/10/27 12:15:47 by ddinaut          ###   ########.fr       */
+/*   Updated: 2018/10/27 17:25:29 by ddinaut          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ static void	print_symbol_x32(t_symbol *symbol)
 	while (tmp != NULL)
 	{
 		if (tmp->value != 0)
-			ft_printf("%08lx %c %s\n", (uint32_t)tmp->value, tmp->type, tmp->name);
+			ft_printf("%08lx %c %s\n", tmp->value, tmp->type, tmp->name);
 		else
 			ft_printf("%8s %c %s\n", "", tmp->type, tmp->name);
 		tmp = tmp->next;
